requirements.txt
# pip install -r requirements.txt
fastapi
uvicorn
typer
psycopg2-binary
python-dotenv
httpx
sqlparse
matplotlib
pandas
plotly
rich
